+++ lib/shared/core/constants/git_commands.dart
@@ -11,8 +11,10 @@ class GitCommands {
   static const List<String> deleteBranch = ["branch", "-D"];
   static const List<String> statusPorcelain = ["status", "--porcelain", "--untracked-files=all"];
   static const List<String> gitAdd = ["add"];
+  static const List<String> gitAddAll = ["add", "-A"];
   static const List<String> gitCommit = ["commit"];
   static const List<String> gitRestoreStaged = ["restore", "--staged"];
+  static const List<String> gitUnstageAll = ["reset"];
   static const List<String> commitsAheadCount = ["rev-list", "--count", "@{u}..HEAD"];
   static const List<String> gitPush = ["push"];
   static const List<String> remoteVerbose = ["remote", "-v"];